# Conflicts:
#	engine-client/pom.xml
#	engine-common/pom.xml
#	engine-common/src/main/java/com/dtstack/engine/common/env/EnvironmentContext.java
#	engine-master/engine-alert/src/main/java/com/dtstack/engine/alert/client/AlertGateFacade.java
#	engine-master/engine-alert/src/main/java/com/dtstack/engine/alert/factory/customize/CustomizeApiService.java
#	engine-master/engine-alert/src/main/java/com/dtstack/engine/alert/send/customize/TencentCloudCustomizeSender.java

+++ engine-master/engine-master-server/src/main/java/com/dtstack/engine/master/impl/TenantService.java
@@ -150,8 +150,8 @@ public class TenantService {
             }
 
             engineTenantVO.setQueue(queue.getQueuePath());
-            engineTenantVO.setMaxCapacity(NumberUtils.toInt(queue.getMaxCapacity(),0) * 100 + "%");
-            engineTenantVO.setMinCapacity(NumberUtils.toInt(queue.getCapacity(),0) * 100 + "%");
+            engineTenantVO.setMaxCapacity(NumberUtils.toDouble(queue.getMaxCapacity(),0) * 100 + "%");
+            engineTenantVO.setMinCapacity(NumberUtils.toDouble(queue.getCapacity(),0) * 100 + "%");
         }
     }
